+++ packages/cli/src/commands/login/future.ts
@@ -38,9 +38,12 @@ export interface DeviceCodeTokens {
  */
 export async function performDeviceCodeFlow(
   client: Client,
-  options?: { teamId?: string }
+  options?: { teamId?: string; refreshToken?: string; acrValues?: string }
 ): Promise<DeviceCodeTokens | null> {
-  const deviceAuthorizationResponse = await deviceAuthorizationRequest();
+  const deviceAuthorizationResponse = await deviceAuthorizationRequest({
+    refresh_token: options?.refreshToken,
+    acr_values: options?.acrValues,
+  });
 
   o.debug(
     `'Device Authorization response:', ${await deviceAuthorizationResponse.clone().text()}`
@@ -223,7 +226,7 @@ export async function login(
     await updateCurrentTeamAfterLogin(client);
   }
 
-  client.writeToAuthConfigFile();
+  client.persistAuthConfig();
   client.writeToConfigFile();
 
   o.debug(`Saved credentials in "${hp(client.getGlobalPathConfig())}"`);