+++ packages/cli/src/util/dev/queue-broker.ts
@@ -3,7 +3,7 @@
 import ms from 'ms';
 import { randomBytes } from 'crypto';
 import nodeFetch from 'node-fetch';
-import type { Service } from '@vercel/fs-detectors';
+import type { ExperimentalService } from '@vercel/fs-detectors';
 import {
   getServiceQueueTopicConfigs,
   isQueueBackedService,
@@ -79,7 +79,7 @@ export class QueueBroker {
   private tickTimer: ReturnType<typeof setInterval>;
 
   constructor(
-    services: Service[],
+    services: ExperimentalService[],
     private getServiceOrigin: (name: string) => string | null
   ) {
     for (const service of services) {