+++ packages/koa/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @vercel/koa
 
+## 0.1.71
+
+### Patch Changes
+
+- @vercel/node@5.8.12
+
+## 0.1.70
+
+### Patch Changes
+
+- @vercel/node@5.8.11
+
+## 0.1.69
+
+### Patch Changes
+
+- @vercel/node@5.8.10
+
+## 0.1.68
+
+### Patch Changes
+
+- @vercel/node@5.8.9
+
 ## 0.1.67
 
 ### Patch Changes